requirements.txt
accelerate==0.29.3
aiohttp==3.9.5
aiosignal==1.3.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cmake==3.29.2
datasets==2.19.1
decorator==5.1.1
dill==0.3.8
einops==0.7.0
exceptiongroup==1.2.1
executing==2.0.1
filelock==3.13.4
frozenlist==1.4.1
fsspec==2024.3.1
huggingface-hub==0.22.2
idna==3.7
# ipython==8.24.0
jedi==0.19.1
Jinja2==3.1.3
jiwer==3.0.4
joblib==1.4.0
lazy_loader==0.4
librosa==0.10.1
lit==18.1.4
# llvmlite==0.42.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
# networkx==3.3
ninja==1.11.1.1
# numba==0.59.1
numpy
openai-whisper==20231117
opencc-python-reimplemented==0.1.7
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.1
pooch==1.8.1
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.0.0
pyarrow-hotfix==0.6
pycparser==2.22
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.9.0
regex==2024.4.16
requests==2.31.0
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
sentencepiece==0.2.0
six==1.16.0
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.3.7
stack-data==0.6.3
sympy==1.12
threadpoolctl==3.4.0
tiktoken==0.6.0
tokenizers==0.19.1
torch
tqdm==4.66.2
traitlets==5.14.3
transformers==4.40.1
triton==2.1.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
xxhash==3.4.1
yarl==1.9.4
